# src/app/api/keys/route.ts
import {NextResponse} from 'next/server';
import {headers} from 'next/headers';
import {clearAllProxyKeys, createProxyKeyForIp} from './service';
import {HttpError} from '@/lib/httpError';
import {extractClientIp} from '@/lib/utils';
import {CreateProxyKeyResponse} from '@/lib/types';

import {createRateLimiter} from '@/lib/rateLimit';

const limiter = createRateLimiter({
  60_000: 3,            // per minute
  21_600_000: 10,       // per 6 hours
  86_400_000: 30,       // per 24 hours
});

export async function POST() {
  const headersList = await headers();
  const ip = extractClientIp(headersList);

  if (ip) {
    const rl = limiter.check(ip);
    if (!rl.allowed) {
      const retryAfterSec = Math.max(1, Math.ceil((rl.resetAt - Date.now()) / 1000));
      return NextResponse.json(
        {message: 'Too many requests'},
        {
          status: 429,
          headers: {
            'Retry-After': String(retryAfterSec),
            'X-RateLimit-Limit': String(rl.limit),
            'X-RateLimit-Remaining': String(rl.remaining),
            'X-RateLimit-Reset': String(Math.floor(rl.resetAt / 1000)),
          },
        },
      );
    }
  }

  try {
    const proxyKey = await createProxyKeyForIp(ip);
    return NextResponse.json({uuid: proxyKey.uuid, accessUrl: proxyKey.accessUrl, createdAt: proxyKey.createdAt} satisfies CreateProxyKeyResponse, {status: 200});
  } catch (e) {
    const status = (e instanceof HttpError) ? e.status : 500;
    const message = (e instanceof HttpError) ? e?.message : 'Internal Server Error';
    return NextResponse.json({message}, {status});
  }
}


export async function DELETE() {
  const headersList = await headers();
  const apiKey = headersList.get('x-api-key');
  const expectedApiKey = process.env.API_KEY;

  if (!apiKey || apiKey !== expectedApiKey) {
    return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
  }

  await clearAllProxyKeys();

  return NextResponse.json({ success: 1 }, { status: 200 });
}